Backend/requirements.txt
beautifulsoup4==4.9
flask
flask-cors
google==3.0.0
langchain
nltk
openai
requests
scikit-learn
youtube-dl==2021.12.17
youtube-transcript-api==0.6.1
translators
selenium
python-docx
python-pptx
Document
SpeechRecognition
pydub
Spire.Presentation
langchain_community
moviepy
python-dotenv
PyJWT
pyunsplash
pytube
pymongo
